server/requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
pydantic==2.12.3
python-dotenv==1.2.1
python-multipart==0.0.20
email-validator==2.3.0
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==5.0.0
sendgrid==6.12.5
